refactor: extract translate finalization
